backend/requirements.txt
fastapi==0.109.1
uvicorn[standard]==0.27.0
sqlalchemy[asyncio]==2.0.25
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.6
pydantic[email]==2.6.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.22
httpx==0.26.0
ccxt==4.2.15
pandas==2.2.0
numpy==1.26.3
pyotp==2.9.0
qrcode[pil]==7.4.2
pillow==12.1.1
cryptography==46.0.5
pytest==7.4.4
pytest-asyncio==0.23.4
aiofiles==23.2.1
